args: use new function find_path instead of readlink
Since we set -e, the script can exit immediately if a pipeline, a
list, or a compound command returns a non-zero status.

When we parse arguments, if we check a path with readlink the program
will exit if the path does not exist.
That is not expected, we want at least display an error and the usage.

By using the new function find_path, we have the expected behavior.

Signed-off-by: Julien Masson <jmasson@baylibre.com>