Don't exit on usage and use new function error_exit
The usage function should not be responsible of exiting the program.
This function should ONLY print usage informations.
Following the cases we might want to have an exit status of 0 or 1.

A new function has been added error_exit and do in this order:
1. display an error message
2. display usage
3. exit 1

error_exit is used when we check program arguments.

Signed-off-by: Julien Masson <jmasson@baylibre.com>