factory: add RPMB support
The corresponding config has been added to the factory
defconfig_fragment in order to build uboot with RPMB support.

OP-TEE is also built with the RPMB flags in order to make RPMB
read/write request to REE (U-Boot or Linux).

WARNING:
- CFG_RPMB_WRITE_KEY flag:
This flag configure the RPMB if this one is not initialized yet.
This operation is performed one time only, once the RPMB is
initialized with the key we cannot override it anymore.

Signed-off-by: Julien Masson <jmasson@baylibre.com>