Fix meson build warnings
The following fields are deprecated and has been removed from
meson.cross: c_args, c_link_args, cpp_args and cpp_link_args

It fix the following warnings:
DEPRECATION: c_args in the [properties] section of the machine file is deprecated, use the [built-in options] section.
DEPRECATION: c_link_args in the [properties] section of the machine file is deprecated, use the [built-in options] section.
DEPRECATION: cpp_args in the [properties] section of the machine file is deprecated, use the [built-in options] section.
DEPRECATION: cpp_link_args in the [properties] section of the machine file is deprecated, use the [built-in options] section.

Signed-off-by: Julien Masson <jmasson@baylibre.com>