libdram: don't use conditional expression end of function
In clean_libdram function, if the folder is not present we return an
error from the function since we are using conditional expression at
the end of the function and the program exit.

This behavior is unwanted, we should not exit from the program.
By using an if statement, the program execute as expected.

Signed-off-by: Julien Masson <jmasson@baylibre.com>