{setup,release}_android: invoke repo with --color=never
On environments with git config color.ui = always, we can't get the repo
branch because the perl oneliner does not match anything.

In fact, the revision will be '?[m?[33mmain?[m' instead of 'main'.

This can be fixed in the environment with:
$ git config --global color.ui never

Since we don't want to force developers to change their gitconfig, pass
--color to "repo info"

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>