commit-binaries: add initial boilerplate code
commit-binaries is intended to replace the commit logic from
release_android.sh into a dedicated file so that it can be reused by
other components, such as the Android Common Kernel.

This adds initial boilerplate code for it and the usage() function which
describes how the script should be used.

It will be implemented in following commits by moving in
code from release_android.sh.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>